+++ src/bindings/python/nixl-meta/nixl_ep/__init__.py
@@ -19,7 +19,7 @@ import importlib
 import sys
 from typing import TYPE_CHECKING
 
-from nixl_utils import detect_cuda_major
+from nixl_meta_utils import detect_cuda_major
 
 
 def _load_ep_module() -> str:
