+++ src/bindings/python/nixl-meta/pyproject.toml.in
@@ -37,4 +37,4 @@ cu13 = ["nixl-cu13==@VERSION@"]
 
 [tool.setuptools]
 packages = ["nixl", "nixl_ep"]
-py-modules = ["nixl_utils"]
+py-modules = ["nixl_meta_utils"]
